+++ packages/plugin-packages/mashroom-portal/src/frontend/portal-client/js/MashroomPortalAppServiceImpl.js
@@ -367,30 +367,27 @@ export default class MashroomPortalAppServiceImpl implements MashroomPortalAppSe
         return this._restService.get(path);
     }
 
-    _loadResources(loadedPortalApp: LoadedPortalAppInternal): Promise<any> {
+    _loadResources(loadedPortalApp: LoadedPortalAppInternal): Promise<void> {
         const appSetup = loadedPortalApp.appSetup;
         if (!appSetup) {
             return Promise.reject('appSetup not loaded');
         }
 
         // JavaScript
-        const sharedJSResourcesPromises: Array<Promise<void>> = [];
+        // Load the script sequentially, first the shared ones
+        let loadJsPromise: Promise<void> =  Promise.resolve();
         if (appSetup.sharedResources && appSetup.sharedResources.js) {
-            appSetup.sharedResources.js.forEach((jsResource) =>
-                sharedJSResourcesPromises.push(
-                    this._resourceManager.loadJs(`${appSetup.sharedResourcesBasePath}/js/${jsResource}`, loadedPortalApp))
-            );
+            appSetup.sharedResources.js.forEach((jsResource) => {
+                const promise = this._resourceManager.loadJs(`${appSetup.sharedResourcesBasePath}/js/${jsResource}`, loadedPortalApp);
+                loadJsPromise = loadJsPromise.then(() => promise);
+            });
         }
-        const jsPromises = Promise.all(sharedJSResourcesPromises).then(() => {
-            const jsResourcesPromises: Array<Promise<void>> = [];
-            if (appSetup.resources.js) {
-                appSetup.resources.js.forEach((jsResource) =>
-                    jsResourcesPromises.push(
-                        this._resourceManager.loadJs(`${appSetup.resourcesBasePath}/${jsResource}?v=${appSetup.lastReloadTs}`, loadedPortalApp))
-                );
-            }
-            return Promise.all(jsResourcesPromises);
-        });
+        if (appSetup.resources.js) {
+            appSetup.resources.js.forEach((jsResource) => {
+                const promise = this._resourceManager.loadJs(`${appSetup.resourcesBasePath}/${jsResource}?v=${appSetup.lastReloadTs}`, loadedPortalApp);
+                loadJsPromise = loadJsPromise.then(() => promise);
+            });
+        }
 
         // CSS
         // We don't have to wait for CSS resources before we can start the app
@@ -403,7 +400,7 @@ export default class MashroomPortalAppServiceImpl implements MashroomPortalAppSe
                 this._resourceManager.loadStyle(`${appSetup.resourcesBasePath}/${cssResource}?v=${appSetup.lastReloadTs}`, loadedPortalApp));
         }
 
-        return jsPromises;
+        return loadJsPromise;
     }
 
     _startApp(appId: string, wrapper: HTMLElement, appSetup: MashroomPortalAppSetup, pluginName: string): Promise<?MashroomPortalAppLifecycleHooks> {